Rename app.mjs to nodeConsoleVersion/app.mjs
Moving this to it's own folder. The assistant started out as just a node console script, but for the sake of practicality and convince I moved it to AWS. Leaving this here as it is the simplest way to get results. This will output in the console, and create a HTML file for easier reading.